refactor: separa os hooks criados
