+++ client/src/hooks/useAuth.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { AuthContext } from "../contexts/AuthContext";
+
+export function useAuth() {
+	const context = useContext(AuthContext);
+	if (!context) {
+		throw new Error("useAuth deve ser usado dentro de um AuthProvider");
+	}
+	return context;
+}
