+++ client/src/pages/Login.tsx
@@ -1,11 +1,11 @@
 import { useState } from "react";
-import { useAuth } from "../contexts/AuthContext";
 import { Button } from "../components/Button/index";
 import { InputContainer } from "../components/Containers/InputContainer";
 import { Input } from "../components/Input";
 import type { Login } from "../types/userTypes";
 import clouds from "../assets/clouds.svg";
 import logo from "../assets/logo.png";
+import { useAuth } from "../hooks/useAuth";
 
 interface ErrorMessage {
 	email: string | null;
